+++ packages/blocks/src/bookmark-block/bookmark-block.ts
@@ -1,203 +1,52 @@
-import './components/bookmark-create-modal.js';
-import './components/bookmark-edit-modal.js';
-import './components/bookmark-toolbar.js';
-import './components/loader.js';
+import './components/index.js';
+import '../_common/components/button.js';
 
-import { Slot } from '@blocksuite/global/utils';
 import { BlockElement } from '@blocksuite/lit';
 import { flip, offset } from '@floating-ui/dom';
-import { css, html, nothing } from 'lit';
-import { customElement, query, state } from 'lit/decorators.js';
+import { html } from 'lit';
+import { customElement, property, query } from 'lit/decorators.js';
 import { ref } from 'lit/directives/ref.js';
 
 import { HoverController } from '../_common/components/index.js';
-import { WebIcon16 } from '../_common/icons/text.js';
-import { stopPropagation } from '../_common/utils/event.js';
-import { getThemeMode } from '../_common/utils/query.js';
 import { AffineDragHandleWidget } from '../_common/widgets/drag-handle/index.js';
 import { captureEventTarget } from '../_common/widgets/drag-handle/utils.js';
 import {
   type BookmarkBlockModel,
   BookmarkBlockSchema,
 } from './bookmark-model.js';
+import type { BookmarkCaption } from './components/bookmark-caption.js';
 import type {
   MenuActionCallback,
   ToolbarActionCallback,
 } from './components/config.js';
-import { embedIframeTemplate } from './embed.js';
-import { DefaultBanner } from './images/banners.js';
-import { DarkLoadingBanner, LoadingBanner } from './images/icons.js';
-import { reloadBookmarkBlock } from './utils.js';
+import { toggleBookmarkEditModal } from './components/modal/bookmark-edit-modal.js';
+import { refreshBookmarkUrlData } from './utils.js';
 
 @customElement('affine-bookmark')
 export class BookmarkBlockComponent extends BlockElement<BookmarkBlockModel> {
-  static override styles = css`
-    .affine-bookmark-block-container {
-      width: 100%;
-      margin-top: 18px;
-      margin-bottom: 18px;
-      position: relative;
-    }
-    .affine-bookmark-link {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      box-shadow: var(--affine-shadow-1);
-      border: 3px solid var(--affine-background-secondary-color);
-      border-radius: 12px;
-      padding: 16px 24px;
-      cursor: pointer;
-      text-decoration: none;
-      color: var(--affine-text-primary-color);
-      overflow: hidden;
-      line-height: calc(1em + 4px);
-      position: relative;
-    }
-    .affine-bookmark-banner {
-      position: absolute;
-      right: 24px;
-      bottom: 0;
-      width: 140px;
-      height: 93px;
-      border-radius: 8px 8px 0 0;
-      overflow: hidden;
-    }
-    .affine-bookmark-banner.shadow {
-      box-shadow: var(--affine-shadow-1);
-    }
+  @query('bookmark-caption')
+  captionElement!: BookmarkCaption;
 
-    .affine-bookmark-banner img,
-    .affine-bookmark-banner svg {
-      width: 140px;
-      height: 93px;
-      object-fit: cover;
-    }
-    .affine-bookmark-content-wrapper {
-      width: 100%;
-      flex-grow: 1;
-      overflow: hidden;
-    }
-    .affine-bookmark-title {
-      height: 18px;
-      display: flex;
-      align-items: center;
-      font-size: var(--affine-font-sm);
-      font-weight: 600;
-    }
-    .affine-bookmark-title-content {
-      flex-grow: 1;
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-      margin-left: 8px;
-    }
-    .affine-bookmark-icon {
-      display: flex;
-      align-items: center;
-      width: 18px;
-      height: 18px;
-      color: var(--affine-text-secondary-color);
-      fill: var(--affine-text-secondary-color);
-      flex-shrink: 0;
-    }
-    .affine-bookmark-icon img {
-      width: 100%;
-      height: 100%;
-      object-fit: cover;
-    }
-    .affine-bookmark-description {
-      height: 32px;
-      line-height: 16px;
-      margin-top: 4px;
-      font-size: var(--affine-font-xs);
-
-      display: -webkit-box;
-      word-break: break-all;
-      -webkit-box-orient: vertical;
-      -webkit-line-clamp: 2;
-      overflow: hidden;
-      text-overflow: ellipsis;
-    }
-    .affine-bookmark-url {
-      font-size: var(--affine-font-xs);
-      color: var(--affine-text-secondary-color);
-      overflow: hidden;
-      text-overflow: ellipsis;
-      white-space: nowrap;
-      margin-top: 2px;
-    }
-    .affine-bookmark-caption {
-      width: 100%;
-      font-size: var(--affine-font-sm);
-      outline: none;
-      border: 0;
-      font-family: inherit;
-      text-align: center;
-      color: var(--affine-icon-color);
-      display: none;
-      background: var(--affine-background-primary-color);
-    }
-    .affine-bookmark-caption::placeholder {
-      color: var(--affine-placeholder-color);
-    }
-    .affine-bookmark-caption.caption-show {
-      display: inline-block;
-    }
-    .affine-bookmark-loading {
-      width: 100%;
-      height: 112px;
-      padding: 16px 24px;
-      display: flex;
-      justify-content: space-between;
-      box-shadow: var(--affine-shadow-1);
-      background: var(--affine-hover-color);
-      border: 3px solid var(--affine-background-secondary-color);
-      color: var(--affine-placeholder-color);
-      border-radius: 12px;
-    }
-
-    .affine-bookmark-embed-frame {
-      grid-area: embed;
-      width: 100%;
-      margin-bottom: 20px;
-      border-radius: 8px;
-      overflow: hidden;
-    }
-  `;
-
-  slots = {
-    openInitialModal: new Slot(),
-  };
-
-  @query('input.affine-bookmark-caption')
-  _input!: HTMLInputElement;
-
-  @state()
-  private _showCreateModal = false;
-
-  @state()
-  private _showEditModal = false;
-
-  @state()
-  private _caption!: string;
-
-  @state()
-  private _isLoading = false;
-  @state()
-  private _isIconError = false;
-  @state()
-  private _isImageError = false;
+  @property({ attribute: false })
+  loading = false;
 
   private _optionsAbortController?: AbortController;
-
-  set loading(value: boolean) {
-    this._isLoading = value;
-  }
-
-  get loading() {
-    return this._isLoading;
-  }
-
+  private _onToolbarSelected: ToolbarActionCallback & MenuActionCallback =
+    type => {
+      if (type === 'caption') {
+        this.captionElement.display = true;
+        this.captionElement.updateComplete.then(() => {
+          this.captionElement.input.focus();
+        });
+      }
+      if (type === 'edit') {
+        toggleBookmarkEditModal(this.model);
+      }
+      if (type === 'reload') {
+        refreshBookmarkUrlData(this);
+      }
+      this._optionsAbortController?.abort();
+    };
   private _whenHover = new HoverController(this, ({ abortController }) => {
     this._optionsAbortController = abortController;
     return {
@@ -221,29 +70,6 @@ export class BookmarkBlockComponent extends BlockElement<BookmarkBlockModel> {
       },
     };
   });
-
-  override firstUpdated() {
-    this.model.propsUpdated.on(() => this.requestUpdate());
-
-    this.updateComplete.then(() => {
-      this._caption = this.model?.caption ?? '';
-
-      if (this._caption) {
-        // Caption input should be toggled manually.
-        // However, it will be lost if the caption is deleted into empty state.
-        this._input.classList.add('caption-show');
-      }
-    });
-  }
-
-  override connectedCallback() {
-    super.connectedCallback();
-    reloadBookmarkBlock(this.model, this);
-    this.slots.openInitialModal.on(() => {
-      this._showCreateModal = true;
-    });
-    this._registerDragHandleOption();
-  }
 
   private _registerDragHandleOption = () => {
     this._disposables.add(
@@ -269,180 +95,25 @@ export class BookmarkBlockComponent extends BlockElement<BookmarkBlockModel> {
     );
   };
 
-  private _onInputChange() {
-    this._caption = this._input.value;
-    this.model.page.updateBlock(this.model, { caption: this._caption });
-  }
-
-  private _onInputBlur() {
-    if (!this._caption) {
-      this._input.classList.remove('caption-show');
-    }
-  }
-
-  private _onCardClick() {
-    const selectionManager = this.root.selection;
-    const blockSelection = selectionManager.getInstance('block', {
-      path: this.path,
-    });
-    selectionManager.setGroup('note', [blockSelection]);
-  }
-
-  private _onCardDbClick() {
-    let link = this.model.url;
-
-    if (!link.match(/^[a-zA-Z]+:\/\//)) {
-      link = 'https://' + link;
-    }
-    window.open(link, '_blank');
-  }
-
-  private _onIconError() {
-    this._isIconError = true;
-  }
-  private _onImageError() {
-    this._isImageError = true;
-  }
-
-  private _onToolbarSelected: ToolbarActionCallback & MenuActionCallback =
-    type => {
-      if (type === 'caption') {
-        this._input.classList.add('caption-show');
-        requestAnimationFrame(() => {
-          this._input.focus();
-        });
-      }
-
-      if (type === 'edit') {
-        this._showEditModal = true;
-      }
-
-      if (type === 'reload') {
-        this._isImageError = false;
-        this._isIconError = false;
-      }
-      this._optionsAbortController?.abort();
-    };
-
-  private _linkCard() {
-    const { url, bookmarkTitle, description, icon, image } = this.model;
-
-    const isEmbed = this.model.type === 'embed';
-    const titleIcon =
-      icon && !this._isIconError
-        ? html`<img src="${icon}" alt="icon" @error="${this._onIconError}" />`
-        : WebIcon16;
-
-    const bannerImage = isEmbed
-      ? nothing
-      : html`<div class="affine-bookmark-banner ${image ? 'shadow' : ''}">
-          ${image && !this._isImageError
-            ? html`<img
-                src="${image}"
-                alt="image"
-                @error="${this._onImageError}"
-              />`
-            : DefaultBanner}
-        </div>`;
-
-    return html`<div
-      class="affine-bookmark-link"
-      style="${isEmbed
-        ? nothing
-        : 'background: var(--affine-card-background-blue);'}"
-      @click=${this._onCardClick}
-      @dblclick=${this._onCardDbClick}
-    >
-      ${isEmbed
-        ? html`<div class="affine-bookmark-embed-frame">
-            ${embedIframeTemplate(url)}
-          </div>`
-        : nothing}
-
-      <div
-        class="affine-bookmark-content-wrapper"
-        style="${isEmbed ? nothing : 'padding-right: 165px;'}"
-      >
-        <div class="affine-bookmark-title">
-          <div class="affine-bookmark-icon">${titleIcon}</div>
-          <div class="affine-bookmark-title-content">
-            ${bookmarkTitle || 'Bookmark'}
-          </div>
-        </div>
-
-        <div class="affine-bookmark-description">${description || url}</div>
-        <div class="affine-bookmark-url">${url}</div>
-      </div>
-
-      ${bannerImage}
-    </div>`;
+  override connectedCallback() {
+    super.connectedCallback();
+    refreshBookmarkUrlData(this);
+    this._registerDragHandleOption();
   }
 
   override render() {
-    const { url } = this.model;
-    const mode = getThemeMode();
-
-    const createModal = this._showCreateModal
-      ? html`<bookmark-create-modal
-          .model=${this.model}
-          .onCancel=${() => {
-            this._showCreateModal = false;
-          }}
-          .onConfirm=${() => {
-            reloadBookmarkBlock(this.model, this, true);
-            this._showCreateModal = false;
-          }}
-        ></bookmark-create-modal>`
-      : nothing;
-    const editModal = this._showEditModal
-      ? html`<bookmark-edit-modal
-          .model=${this.model}
-          .onCancel=${() => {
-            this._showEditModal = false;
-          }}
-          .onConfirm=${() => {
-            this._showEditModal = false;
-          }}
-        ></bookmark-edit-modal>`
-      : nothing;
-
-    const loading = this._isLoading
-      ? html`<div
-          class="affine-bookmark-loading ${mode === 'light' ? '' : 'dark'}"
-        >
-          <div class="affine-bookmark-title">
-            <bookmark-loader
-              .size=${'15px'}
-              .color=${'var(--affine-primary-color)'}
-            ></bookmark-loader>
-            <div class="affine-bookmark-title-content">Loading...</div>
-          </div>
-          <div class="affine-bookmark-banner">
-            ${mode === 'light' ? LoadingBanner : DarkLoadingBanner}
-          </div>
-        </div>`
-      : nothing;
-
-    if (!url) {
-      return createModal;
-    }
-
     return html`
-      ${editModal}
       <div
         ${ref(this._whenHover.setReference)}
         class="affine-bookmark-block-container"
       >
-        ${this._isLoading ? loading : this._linkCard()}
-        <input
-          .disabled=${this.model.page.readonly}
-          placeholder="Write a caption"
-          class="affine-bookmark-caption"
-          value=${this._caption}
-          @input=${this._onInputChange}
-          @blur=${this._onInputBlur}
-          @pointerdown=${stopPropagation}
-        />
+        ${this.loading
+          ? html`<bookmark-loading></bookmark-loading>`
+          : html`<bookmark-card .bookmark=${this}></bookmark-card>`}
+        <bookmark-caption
+          .bookmark=${this}
+          .display=${this.model.caption && this.model.caption.length > 0}
+        ></bookmark-caption>
         ${this.selected?.is('block')
           ? html`<affine-block-selection></affine-block-selection>`
           : null}
