+++ packages/blocks/src/bookmark-block/components/modal/index.ts
@@ -1,0 +1,2 @@
+export * from './bookmark-create-modal.js';
+export * from './bookmark-edit-modal.js';
